remove registry
