+++ src/main/resources/answers.properties
@@ -27,8 +27,5 @@ answers.start.description=Just start command
 answers.start.message=Hi, i'm a Blamer!
 answers.token.command=/token
 answers.token.description=Set GitHub token to get updates
-answers.token.message=You need to provide me GitHub token with notification access! Send it with /registry your_token
-answers.registry.command=/registry
-answers.registry.description=Token authentication
 answers.unsubscribe.command=/unsubscribe
 answers.unsubscribe.description=Unsubscribes you
