+++ src/test/resources/answers.properties
@@ -27,8 +27,5 @@ answers.start.description=test start description
 answers.start.message=test start message
 answers.token.command=/token
 answers.token.description=test token description
-answers.token.message=test token message
-answers.registry.command=/registry
-answers.registry.description=test registry description
 answers.unsubscribe.command=/unsubscribe
 answers.unsubscribe.description=Unsubscribes you
